Update README.md
This project implements an autonomous line-following robot capable of detecting and following a black path on a white surface. It uses two IR reflective sensors to detect the line and an Arduino Uno to process sensor inputs and control the motors accordingly.

The robot is suitable for beginners in robotics and helps build understanding of embedded systems, motor drivers, and real-time sensor feedback control.